+++ tests/PhpWord/Reader/Word2007/StyleTest.php
@@ -20,6 +20,7 @@ namespace PhpOffice\PhpWord\Reader\Word2007;
 use PhpOffice\PhpWord\AbstractTestReader;
 use PhpOffice\PhpWord\SimpleType\TblWidth;
 use PhpOffice\PhpWord\Style\Table;
+use PhpOffice\PhpWord\Style\TablePosition;
 
 /**
  * Test class for PhpOffice\PhpWord\Reader\Word2007\Styles
@@ -61,10 +62,42 @@ class StyleTest extends AbstractTestReader
         $elements = $phpWord->getSection(0)->getElements();
         $this->assertInstanceOf('PhpOffice\PhpWord\Element\Table', $elements[0]);
         $this->assertInstanceOf('PhpOffice\PhpWord\Style\Table', $elements[0]->getStyle());
-        $this->assertEquals(TblWidth::AUTO, $elements[0]->getStyle()->getUnit());
         /** @var \PhpOffice\PhpWord\Style\Table $tableStyle */
         $tableStyle = $elements[0]->getStyle();
+        $this->assertEquals(TblWidth::AUTO, $tableStyle->getUnit());
         $this->assertEquals(10.5, $tableStyle->getCellSpacing());
+    }
+
+    /**
+     * Test reading of table position
+     */
+    public function testReadTablePosition()
+    {
+        $documentXml = '<w:tbl>
+            <w:tblPr>
+                <w:tblpPr w:leftFromText="10" w:rightFromText="20" w:topFromText="30" w:bottomFromText="40" w:vertAnchor="page" w:horzAnchor="margin" w:tblpXSpec="center" w:tblpX="50" w:tblpYSpec="top" w:tblpY="60"/>
+            </w:tblPr>
+        </w:tbl>';
+
+        $phpWord = $this->getDocumentFromString(array('document' => $documentXml));
+
+        $elements = $phpWord->getSection(0)->getElements();
+        $this->assertInstanceOf('PhpOffice\PhpWord\Element\Table', $elements[0]);
+        $this->assertInstanceOf('PhpOffice\PhpWord\Style\Table', $elements[0]->getStyle());
+        $this->assertNotNull($elements[0]->getStyle()->getPosition());
+        $this->assertInstanceOf('PhpOffice\PhpWord\Style\TablePosition', $elements[0]->getStyle()->getPosition());
+        /** @var \PhpOffice\PhpWord\Style\TablePosition $tableStyle */
+        $tableStyle = $elements[0]->getStyle()->getPosition();
+        $this->assertEquals(10, $tableStyle->getLeftFromText());
+        $this->assertEquals(20, $tableStyle->getRightFromText());
+        $this->assertEquals(30, $tableStyle->getTopFromText());
+        $this->assertEquals(40, $tableStyle->getBottomFromText());
+        $this->assertEquals(TablePosition::VANCHOR_PAGE, $tableStyle->getVertAnchor());
+        $this->assertEquals(TablePosition::HANCHOR_MARGIN, $tableStyle->getHorzAnchor());
+        $this->assertEquals(TablePosition::XALIGN_CENTER, $tableStyle->getTblpXSpec());
+        $this->assertEquals(50, $tableStyle->getTblpX());
+        $this->assertEquals(TablePosition::YALIGN_TOP, $tableStyle->getTblpYSpec());
+        $this->assertEquals(60, $tableStyle->getTblpY());
     }
 
     /**